@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 MIGRATION_CLASS =
   if ActiveRecord::VERSION::MAJOR >= 5
     ActiveRecord::Migration["#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}"]
@@ -8,14 +10,15 @@ MIGRATION_CLASS =
 class CreateFriendlyIdSlugs < MIGRATION_CLASS
   def change
     create_table :friendly_id_slugs do |t|
-      t.string   :slug,           :null => false
-      t.integer  :sluggable_id,   :null => false
-      t.string   :sluggable_type, :limit => 50
+      t.string   :slug,           null: false
+      t.integer  :sluggable_id,   null: false
+      t.string   :sluggable_type, limit: 50
       t.string   :scope
       t.datetime :created_at
     end
-    add_index :friendly_id_slugs, [:sluggable_type, :sluggable_id]
-    add_index :friendly_id_slugs, [:slug, :sluggable_type], length: { slug: 140, sluggable_type: 50 }
-    add_index :friendly_id_slugs, [:slug, :sluggable_type, :scope], length: { slug: 70, sluggable_type: 50, scope: 70 }, unique: true
+    add_index :friendly_id_slugs, %i[sluggable_type sluggable_id]
+    add_index :friendly_id_slugs, %i[slug sluggable_type], length: { slug: 140, sluggable_type: 50 }
+    add_index :friendly_id_slugs, %i[slug sluggable_type scope],
+              length: { slug: 70, sluggable_type: 50, scope: 70 }, unique: true
   end
 end
