@@ -1,15 +1,15 @@
-module ApplicationHelper
+# frozen_string_literal: true
 
-  def flash_messages(opts = {})
+module ApplicationHelper
+  def flash_messages(_opts = {})
     flash.each do |msg_type, msg|
       concat(
         content_tag(:div, msg, class: "alert alert-#{msg_type} alert-dismissable fade-in") do
-          concat content_tag(:button, 'x', class: "close", data: { dismiss: 'alert' })
+          concat content_tag(:button, 'x', class: 'close', data: { dismiss: 'alert' })
           concat msg.html_safe
         end
       )
     end
     nil
   end
-  
 end
