@@ -1,16 +1,18 @@
+# frozen_string_literal: true
+
 module PostsHelper
   def medium_url(post)
     if post.medium_url.present?
-      if url = URI.decode(post.medium_url).match(/(?<=shih\/).*/)
+      if url = URI.decode(post.medium_url).match(%r{(?<=shih/).*})
         [url[0].truncate(20), post.medium_url]
       end
     else
-      ["NO DATA"]
+      ['NO DATA']
     end
   end
 
   def render_section_subheading
-    p = /[^\/]*$/.match(request.path)[0].upcase
+    p = %r{[^/]*$}.match(request.path)[0].upcase
     p.present? ? p : nil
   end
 
