@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class PostsControllerTest < ActionDispatch::IntegrationTest
@@ -9,40 +11,44 @@ class PostsControllerTest < ActionDispatch::IntegrationTest
     Rails.cache.clear
   end
 
-  test "should get index" do
+  test 'should get index' do
     get posts_url
     assert_response :success
   end
 
-  test "should get new" do
+  test 'should get new' do
     get new_post_url
     assert_response :success
   end
 
-  test "should create post" do
+  test 'should create post' do
     assert_difference('Post.count') do
-      post posts_url, params: { post: { body: @post.body, description: @post.description, status: @post.status, title: @post.title } }
+      post posts_url,
+           params: { post: { body: @post.body, description: @post.description, status: @post.status,
+                             title: @post.title } }
     end
 
     assert_redirected_to list_posts_url
   end
 
-  test "should show post" do
+  test 'should show post' do
     get post_url(@post)
     assert_response :success
   end
 
-  test "should get edit" do
+  test 'should get edit' do
     get edit_post_url(@post)
     assert_response :success
   end
 
-  test "should update post" do
-    patch post_url(@post), params: { post: { body: @post.body, description: @post.description, status: @post.status, title: @post.title } }
+  test 'should update post' do
+    patch post_url(@post),
+          params: { post: { body: @post.body, description: @post.description, status: @post.status,
+                            title: @post.title } }
     assert_redirected_to post_url(@post)
   end
 
-  test "should destroy post" do
+  test 'should destroy post' do
     assert_difference('Post.count', -1) do
       delete post_url(@post)
     end
